Moved db/models to __init__.py
