--- db/models/base.py
@@ -1,24 +1,0 @@
-"""DB Document base models."""
-
-from typing import Optional
-
-from datetime import datetime
-
-from pydantic import BaseModel
-from pydantic import Field
-
-
-class DocumentBaseModel(BaseModel):
-    """DocumentBaseModel."""
-
-    id: str = Field(None, alias="_id")
-    inserted_at: Optional[datetime]
-    updated_at: Optional[datetime]
-
-    def dict(self, **kwargs) -> dict:
-        values = super().dict(**kwargs)
-        if "id" in values and values["id"]:
-            values["_id"] = values["id"]
-        if "exclude" in kwargs and "_id" in kwargs["exclude"]:
-            values.pop("_id")
-        return values
